impl `ff::{Field, PrimeField, FromBytesUniform}` as macro
