+++ Cargo.toml
@@ -40,7 +40,6 @@ paste = "1.0.11"
 serde = { version = "1.0", default-features = false, optional = true }
 serde_arrays = { version = "0.1.0", optional = true }
 hex = { version = "0.4", optional = true, default-features = false, features = ["alloc", "serde"] }
-blake2b_simd = "1"
 rayon = "1.8"
 unroll = "0.1.5"
 blake2 = "0.10.6"
