+++ src/bn256/fr.rs
@@ -1,7 +1,10 @@
 #[cfg(feature = "asm")]
 use crate::bn256::assembly::field_arithmetic_asm;
 #[cfg(not(feature = "asm"))]
-use crate::{arithmetic::macx, field_arithmetic, field_specific};
+use crate::{
+    arithmetic::{bigint_geq, macx},
+    field_arithmetic_4, field_specific_4,
+};
 
 #[cfg(feature = "bn256-table")]
 #[rustfmt::skip]
@@ -18,33 +21,23 @@ pub use table::FR_TABLE;
 #[cfg(not(feature = "bn256-table"))]
 use crate::impl_from_u64;
 
-use crate::arithmetic::{adc, bigint_geq, mac, sbb};
-use crate::extend_field_legendre;
-use crate::ff::{FromUniformBytes, PrimeField, WithSmallOrderMulGroup};
-use crate::{
-    field_bits, field_common, impl_add_binop_specify_output, impl_binops_additive,
-    impl_binops_additive_specify_output, impl_binops_multiplicative,
-    impl_binops_multiplicative_mixed, impl_sub_binop_specify_output, impl_sum_prod,
-};
+use crate::arithmetic::{adc, mac, sbb};
+use crate::ff::{Field, FromUniformBytes, PrimeField, WithSmallOrderMulGroup};
+use crate::serde::SerdeObject;
 use core::convert::TryInto;
 use core::fmt;
 use core::ops::{Add, Mul, Neg, Sub};
 use rand::RngCore;
 use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
 
-/// This represents an element of $\mathbb{F}_r$ where
-///
-/// `r = 0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001`
-///
-/// is the scalar field of the BN254 curve.
-// The internal representation of this type is four 64-bit unsigned
-// integers in little-endian order. `Fr` values are always in
-// Montgomery form; i.e., Fr(a) = aR mod r, with R = 2^256.
-#[derive(Clone, Copy, PartialEq, Eq, Hash)]
-pub struct Fr(pub(crate) [u64; 4]);
+// Number of 64 bit limbs to represent the field element
+pub(crate) const NUM_BITS: u32 = 254;
 
-#[cfg(feature = "derive_serde")]
-crate::serialize_deserialize_32_byte_primefield!(Fr);
+// Inverter constant
+const BYIL: usize = 6;
+
+// Jabobi constant
+const JACOBI_L: usize = 5;
 
 /// Constant representing the modulus
 /// r = 0x30644e72e131a029b85045b68181585d2833e84879b9709143e1f593f0000001
@@ -102,8 +95,9 @@ const R3: Fr = Fr([
 
 /// `GENERATOR = 7 mod r` is a generator of the `r - 1` order multiplicative
 /// subgroup, or in other words a primitive root of the field.
-const GENERATOR: Fr = Fr::from_raw([0x07, 0x00, 0x00, 0x00]);
+const MULTIPLICATIVE_GENERATOR: Fr = Fr::from_raw([0x07, 0x00, 0x00, 0x00]);
 
+/// Size of the 2-adic sub-group of the field.
 const S: u32 = 28;
 
 /// GENERATOR^t where t * 2^s + 1 = r
@@ -117,20 +111,20 @@ const ROOT_OF_UNITY: Fr = Fr::from_raw([
     0x03ddb9f5166d18b7,
 ]);
 
-/// 1 / 2 mod r
-const TWO_INV: Fr = Fr::from_raw([
-    0xa1f0fac9f8000001,
-    0x9419f4243cdcb848,
-    0xdc2822db40c0ac2e,
-    0x183227397098d014,
-]);
-
 /// 1 / ROOT_OF_UNITY mod r
 const ROOT_OF_UNITY_INV: Fr = Fr::from_raw([
     0x0ed3e50a414e6dba,
     0xb22625f59115aba7,
     0x1bbe587180f34361,
     0x048127174daabc26,
+]);
+
+/// 1 / 2 mod r
+const TWO_INV: Fr = Fr::from_raw([
+    0xa1f0fac9f8000001,
+    0x9419f4243cdcb848,
+    0xdc2822db40c0ac2e,
+    0x183227397098d014,
 ]);
 
 /// GENERATOR^{2^s} where t * 2^s + 1 = r with t odd. In other words, this is a t root of unity.
@@ -150,26 +144,36 @@ const ZETA: Fr = Fr::from_raw([
     0x30644e72e131a029,
 ]);
 
+use crate::{
+    const_montgomery_4, extend_field_legendre, field_bits, impl_add_binop_specify_impl,
+    impl_add_binop_specify_output, impl_binops_additive, impl_binops_additive_specify_output,
+    impl_binops_multiplicative, impl_binops_multiplicative_mixed, impl_field,
+    impl_from_uniform_bytes, impl_prime_field, impl_serde_object, impl_sub_binop_specify_output,
+    impl_sum_prod, pow_vartime,
+};
 impl_binops_additive!(Fr, Fr);
 impl_binops_multiplicative!(Fr, Fr);
-field_common!(
-    Fr,
-    MODULUS,
-    INV,
-    MODULUS_STR,
-    TWO_INV,
-    ROOT_OF_UNITY_INV,
-    DELTA,
-    ZETA,
-    R,
-    R2,
-    R3
-);
+impl_add_binop_specify_impl!(Fr);
+impl_field!(Fr, sparse);
+impl_serde_object!(Fr);
+impl_prime_field!(Fr, [u8; 32], le);
 impl_sum_prod!(Fr);
 extend_field_legendre!(Fr);
+impl_from_uniform_bytes!(Fr, 64);
+
+const_montgomery_4!(Fr);
+#[cfg(not(feature = "asm"))]
+field_arithmetic_4!(Fr, sparse);
+#[cfg(feature = "asm")]
+field_arithmetic_asm!(Fr, MODULUS, INV);
+
+#[cfg(target_pointer_width = "64")]
+field_bits!(Fr);
+#[cfg(not(target_pointer_width = "64"))]
+field_bits!(Fr);
 
 #[cfg(not(feature = "bn256-table"))]
-impl_from_u64!(Fr, R2);
+impl_from_u64!(Fr);
 #[cfg(feature = "bn256-table")]
 // A field element is represented in the montgomery form -- this allows for cheap mul_mod operations.
 // The catch is, if we build an Fr element, regardless of its format, we need to perform one big integer multiplication:
@@ -190,54 +194,10 @@ impl From<u64> for Fr {
     }
 }
 
-#[cfg(not(feature = "asm"))]
-field_arithmetic!(Fr, MODULUS, INV, sparse);
-#[cfg(feature = "asm")]
-field_arithmetic_asm!(Fr, MODULUS, INV);
-
-#[cfg(target_pointer_width = "64")]
-field_bits!(Fr, MODULUS);
-#[cfg(not(target_pointer_width = "64"))]
-field_bits!(Fr, MODULUS, MODULUS_LIMBS_32);
+#[cfg(feature = "derive_serde")]
+crate::serialize_deserialize_primefield!(Fr, [u8; 32]);
 
 impl Fr {
-    pub const fn size() -> usize {
-        32
-    }
-}
-
-impl ff::Field for Fr {
-    const ZERO: Self = Self::zero();
-    const ONE: Self = Self::one();
-
-    fn random(mut rng: impl RngCore) -> Self {
-        Self::from_u512([
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-            rng.next_u64(),
-        ])
-    }
-
-    fn double(&self) -> Self {
-        self.double()
-    }
-
-    #[inline(always)]
-    fn square(&self) -> Self {
-        self.square()
-    }
-
-    /// Returns the multiplicative inverse of the
-    /// element. If it is zero, the method fails.
-    fn invert(&self) -> CtOption<Self> {
-        self.invert()
-    }
-
     fn sqrt(&self) -> CtOption<Self> {
         /// `(t - 1) // 2` where t * 2^s + 1 = p with t odd.
         const T_MINUS1_OVER2: [u64; 4] = [
@@ -248,86 +208,6 @@ impl ff::Field for Fr {
         ];
         ff::helpers::sqrt_tonelli_shanks(self, T_MINUS1_OVER2)
     }
-
-    fn sqrt_ratio(num: &Self, div: &Self) -> (Choice, Self) {
-        ff::helpers::sqrt_ratio_generic(num, div)
-    }
-}
-
-impl ff::PrimeField for Fr {
-    type Repr = [u8; 32];
-
-    const NUM_BITS: u32 = 254;
-    const CAPACITY: u32 = 253;
-    const MODULUS: &'static str = MODULUS_STR;
-    const MULTIPLICATIVE_GENERATOR: Self = GENERATOR;
-    const ROOT_OF_UNITY: Self = ROOT_OF_UNITY;
-    const ROOT_OF_UNITY_INV: Self = ROOT_OF_UNITY_INV;
-    const TWO_INV: Self = TWO_INV;
-    const DELTA: Self = DELTA;
-    const S: u32 = S;
-
-    fn from_repr(repr: Self::Repr) -> CtOption<Self> {
-        let mut tmp = Fr([0, 0, 0, 0]);
-
-        tmp.0[0] = u64::from_le_bytes(repr[0..8].try_into().unwrap());
-        tmp.0[1] = u64::from_le_bytes(repr[8..16].try_into().unwrap());
-        tmp.0[2] = u64::from_le_bytes(repr[16..24].try_into().unwrap());
-        tmp.0[3] = u64::from_le_bytes(repr[24..32].try_into().unwrap());
-
-        // Try to subtract the modulus
-        let (_, borrow) = sbb(tmp.0[0], MODULUS.0[0], 0);
-        let (_, borrow) = sbb(tmp.0[1], MODULUS.0[1], borrow);
-        let (_, borrow) = sbb(tmp.0[2], MODULUS.0[2], borrow);
-        let (_, borrow) = sbb(tmp.0[3], MODULUS.0[3], borrow);
-
-        // If the element is smaller than MODULUS then the
-        // subtraction will underflow, producing a borrow value
-        // of 0xffff...ffff. Otherwise, it'll be zero.
-        let is_some = (borrow as u8) & 1;
-
-        // Convert to Montgomery form by computing
-        // (a.R^0 * R^2) / R = a.R
-        tmp *= &R2;
-
-        CtOption::new(tmp, Choice::from(is_some))
-    }
-
-    fn to_repr(&self) -> Self::Repr {
-        let tmp: [u64; 4] = (*self).into();
-        let mut res = [0; 32];
-        res[0..8].copy_from_slice(&tmp[0].to_le_bytes());
-        res[8..16].copy_from_slice(&tmp[1].to_le_bytes());
-        res[16..24].copy_from_slice(&tmp[2].to_le_bytes());
-        res[24..32].copy_from_slice(&tmp[3].to_le_bytes());
-
-        res
-    }
-
-    fn is_odd(&self) -> Choice {
-        Choice::from(self.to_repr()[0] & 1)
-    }
-}
-
-impl FromUniformBytes<64> for Fr {
-    /// Converts a 512-bit little endian integer into
-    /// an `Fr` by reducing by the modulus.
-    fn from_uniform_bytes(bytes: &[u8; 64]) -> Self {
-        Self::from_u512([
-            u64::from_le_bytes(bytes[0..8].try_into().unwrap()),
-            u64::from_le_bytes(bytes[8..16].try_into().unwrap()),
-            u64::from_le_bytes(bytes[16..24].try_into().unwrap()),
-            u64::from_le_bytes(bytes[24..32].try_into().unwrap()),
-            u64::from_le_bytes(bytes[32..40].try_into().unwrap()),
-            u64::from_le_bytes(bytes[40..48].try_into().unwrap()),
-            u64::from_le_bytes(bytes[48..56].try_into().unwrap()),
-            u64::from_le_bytes(bytes[56..64].try_into().unwrap()),
-        ])
-    }
-}
-
-impl WithSmallOrderMulGroup<3> for Fr {
-    const ZETA: Self = ZETA;
 }
 
 #[cfg(test)]
@@ -344,18 +224,4 @@ mod test {
     crate::field_testing_suite!(Fr, "sqrt");
     crate::field_testing_suite!(Fr, "zeta");
     crate::field_testing_suite!(Fr, "from_uniform_bytes", 64);
-
-    #[test]
-    fn bench_fr_from_u16() {
-        use ark_std::{end_timer, start_timer};
-
-        let repeat = 10000000;
-        let mut rng = ark_std::test_rng();
-        let base = (0..repeat).map(|_| (rng.next_u32() % (1 << 16)) as u64);
-
-        let timer = start_timer!(|| format!("generate {repeat} Bn256 scalar field elements"));
-        let _res: Vec<_> = base.map(Fr::from).collect();
-
-        end_timer!(timer);
-    }
 }
