+++ src/secp256k1/curve.rs
@@ -9,7 +9,6 @@ use core::cmp;
 use core::fmt::Debug;
 use core::iter::Sum;
 use core::ops::{Add, Mul, Neg, Sub};
-use ff::FromUniformBytes;
 use group::cofactor::CofactorGroup;
 use rand::RngCore;
 use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
@@ -164,38 +163,6 @@ impl IsoSecp256k1 {
             Self::SSWU_Z,
             crate::hash_to_curve::Method::SSWU,
         )
-    }
-}
-
-impl FromUniformBytes<48> for Fp {
-    fn from_uniform_bytes(bytes: &[u8; 48]) -> Self {
-        let repr = &mut [0u8; Self::size()];
-
-        (*repr)[0..24].copy_from_slice(&bytes[..24]);
-        let e0 = Fp::from_repr(*repr).unwrap();
-        (*repr)[0..24].copy_from_slice(&bytes[24..]);
-        let e1 = Fp::from_repr(*repr).unwrap();
-
-        // 2^192
-        const SHIFTER: Fp = Fp::from_raw([0, 0, 0, 1]);
-
-        e0 + e1 * SHIFTER
-    }
-}
-
-impl FromUniformBytes<48> for Fq {
-    fn from_uniform_bytes(bytes: &[u8; 48]) -> Self {
-        let repr = &mut [0u8; Self::size()];
-
-        (*repr)[0..24].copy_from_slice(&bytes[..24]);
-        let e0 = Fq::from_repr(*repr).unwrap();
-        (*repr)[0..24].copy_from_slice(&bytes[24..]);
-        let e1 = Fq::from_repr(*repr).unwrap();
-
-        // 2^192
-        const SHIFTER: Fq = Fq::from_raw([0, 0, 0, 1]);
-
-        e0 + e1 * SHIFTER
     }
 }
