+++ src/tests/field.rs
@@ -352,7 +352,7 @@ macro_rules! field_testing_suite {
                 0x59, 0x62, 0xbe, 0x5d, 0x76, 0x3d, 0x31, 0x8d, 0x17, 0xdb, 0x37, 0x32, 0x54, 0x06,
                 0xbc, 0xe5,
             ]);
-            const LIMBS: usize = $field::size() / 8;
+            const LIMBS: usize = $field::SIZE / 8;
             // failure check
             for _ in 0..1000000 {
                 let rand_word = [(); LIMBS].map(|_| rng.next_u64());
