+++ src/tests/mod.rs
@@ -6,17 +6,18 @@ pub mod curve;
 pub mod field;
 
 pub(crate) fn hex_to_bytes(hex: &str) -> Vec<u8> {
-    hex.as_bytes()
+    let bytes = hex.as_bytes().to_vec();
+    bytes
         .chunks(2)
         .map(|chunk| u8::from_str_radix(std::str::from_utf8(chunk).unwrap(), 16).unwrap())
         .collect()
 }
 
 pub(crate) fn hex_to_field<F: PrimeField>(hex: &str) -> F {
-    let bytes = hex_to_bytes(hex);
+    let mut bytes = hex_to_bytes(hex);
+    bytes.reverse();
     let mut repr = F::Repr::default();
-    repr.as_mut().copy_from_slice(&bytes);
-    repr.as_mut().reverse();
+    repr.as_mut()[..bytes.len()].copy_from_slice(&bytes);
     F::from_repr(repr).unwrap()
 }
